requirements.txt
pandas==1.5.3
numpy==1.23.5
pandas-ta==0.3.14b0
polygon-api-client>=1.14.3
xgboost==1.7.5
scikit-learn==1.2.2
joblib==1.2.0

